[jupyter] move log widget to tools.jupyter
