+++ docs/source/jupyter_init.txt
@@ -5,7 +5,7 @@
     from IPython import get_ipython
     ip = get_ipython()
     if ip is not None:
-        ip.run_line_magic('load_ext', 'pymor.discretizers.builtin.gui.jupyter')
+        ip.run_line_magic('load_ext', 'pymor.tools.jupyter')
     %matplotlib inline
 
     import warnings
